+++ stopify-continuations/src/runtime/lazyRuntime.ts
@@ -1,5 +1,6 @@
 import * as common from './abstractRuntime';
 export * from './abstractRuntime';
+import { Result } from '../types';
 
 export class LazyRuntime extends common.Runtime {
   constructor(stackSize: number, restoreFrames: number) {
@@ -33,6 +34,12 @@ export class LazyRuntime extends common.Runtime {
     };
   }
 
+  endTurn(callback: (onDone: (x: Result) => any) => any): never {
+    this.savedStack = [];
+    this.stack = [];
+    throw new common.EndTurn(callback);
+  }
+
   abstractRun(body: () => any): common.RunResult {
     try {
       const v = body();
@@ -45,6 +52,9 @@ export class LazyRuntime extends common.Runtime {
       }
       else if (exn instanceof common.Restore) {
         return { type: 'restore', stack: exn.stack, savedStack: exn.savedStack };
+      }
+      else if (exn instanceof common.EndTurn) {
+        return { type: 'end-turn', callback: exn.callback };
       }
       else {
         return { type: 'exception', value: exn };