+++ stopify/src/runtime/abstractRunner.ts
@@ -1,7 +1,19 @@
 import { RuntimeOpts, AsyncRun } from '../types';
 import { Runtime } from 'stopify-continuations/dist/src/types';
-import { RuntimeWithSuspend } from './suspend';
+import { RuntimeWithSuspend, badResume } from './suspend';
 import { makeEstimator } from './elapsedTimeEstimator';
+import { Result } from 'stopify-continuations/dist/src/types';
+
+enum EventProcessingMode {
+  Running,
+  Paused,
+  Waiting
+}
+
+interface EventHandler {
+  body: () => void;
+  receiver: (x: Result) => void;
+}
 
 export abstract class AbstractRunner implements AsyncRun {
   private continuationsRTS: Runtime;
@@ -11,10 +23,13 @@ export abstract class AbstractRunner implements AsyncRun {
   private onBreakpoint: (line: number) => void = function() { };
   private breakpoints: number[] = [];
   private k: any;
+  // The runtime system starts executing the main body of the program.
+  private eventMode = EventProcessingMode.Running;
+  private eventQueue: EventHandler[] = [];
 
   constructor(private opts: RuntimeOpts) { }
 
-  mayYieldRunning(): boolean {
+  private mayYieldRunning(): boolean {
     const n = this.suspendRTS.rts.linenum;
     if (typeof n !== 'number') {
       return false;
@@ -22,7 +37,7 @@ export abstract class AbstractRunner implements AsyncRun {
     return this.breakpoints.includes(n);
   }
 
-  onYieldRunning() {
+  private onYieldRunning() {
     if (this.mayYieldRunning()) {
       this.onBreakpoint(this.suspendRTS.rts.linenum!);
       return false;
@@ -58,7 +73,9 @@ export abstract class AbstractRunner implements AsyncRun {
    */
   onEnd(): void {
     if (this.continuationsRTS.delimitDepth === 1) {
+      this.eventMode = EventProcessingMode.Waiting;
       this.onDone();
+      this.processQueuedEvents();
     }
   }
 
@@ -79,20 +96,34 @@ export abstract class AbstractRunner implements AsyncRun {
     onBreakpoint?: (line: number) => void): void;
 
   pause(onPaused: (line?: number) => void) {
-    this.suspendRTS.onYield = () => {
+    if (this.eventMode === EventProcessingMode.Paused) {
+      throw new Error('the program is already paused');
+    }
+
+    if (this.eventMode === EventProcessingMode.Waiting) {
+      this.suspendRTS.onYield = function() {
+        throw new Error('Stopify internal error: onYield invoked during pause+wait');
+      }
+      onPaused(); // onYield will not be invoked
+    }
+    else {
       this.suspendRTS.onYield = () => {
-        this.onYield();
-        return true;
+        this.suspendRTS.onYield = () => {
+          this.onYield();
+          return true;
+        };
+        const maybeLine = this.suspendRTS.rts.linenum;
+        if (typeof maybeLine === 'number') {
+          onPaused(maybeLine);
+        }
+        else {
+          onPaused();
+        }
+        return false;
       };
-      const maybeLine = this.suspendRTS.rts.linenum;
-      if (typeof maybeLine === 'number') {
-        onPaused(maybeLine);
-      }
-      else {
-        onPaused();
-      }
-      return false;
-    };
+    }
+
+    this.eventMode = EventProcessingMode.Paused;
   }
 
   setBreakpoints(lines: number[]): void {
@@ -100,12 +131,33 @@ export abstract class AbstractRunner implements AsyncRun {
   }
 
   resume() {
-    this.suspendRTS.mayYield = () => this.mayYieldRunning();
-    this.suspendRTS.onYield = () => this.onYieldRunning();
-    this.suspendRTS.resumeFromCaptured();
+    if (this.eventMode === EventProcessingMode.Waiting) {
+      return;
+    }
+
+    if (this.eventMode === EventProcessingMode.Running) {
+      throw new Error(`invokes .resume() while the program is running`);
+    }
+
+    // Program was paused, but there was no active continuation.
+    if (this.suspendRTS.continuation === badResume) {
+      this.eventMode = EventProcessingMode.Waiting;
+      this.processQueuedEvents();
+    }
+    else {
+      this.eventMode = EventProcessingMode.Running;
+      this.suspendRTS.mayYield = () => this.mayYieldRunning();
+      this.suspendRTS.onYield = () => this.onYieldRunning();
+      this.suspendRTS.resumeFromCaptured();
+    }
   }
 
   step(onStep: (line: number) => void) {
+    if (this.eventMode !== EventProcessingMode.Paused) {
+      throw new Error(`step(onStep) requires the program to be paused`);
+    }
+
+    // NOTE: The program remains Paused while stepping.
     const currentLine = this.suspendRTS.rts.linenum;
     // Yield control if the line number changes.
     const mayYield = () => {
@@ -128,14 +180,42 @@ export abstract class AbstractRunner implements AsyncRun {
   }
 
   pauseImmediate(callback: () => void): void {
+    // NOTE: We don't switch the mode to paused! This is because we don't want
+    // the user to be able to hit "step" at this point.
     return this.continuationsRTS.captureCC((k) => {
-      this.k = k;
-      callback();
+      return this.continuationsRTS.endTurn(onDone => {
+        this.k = { k, onDone }
+        callback();
+      });
     });
   }
 
   continueImmediate(result: any): void {
-    return this.continuationsRTS.runtime(() => this.k(result));
+    const { k, onDone } = this.k;
+    this.k = undefined;
+    return this.continuationsRTS.runtime(() => k(result), (result) => onDone(result));
   }
 
+  processEvent(body: () => void, receiver: (x: Result) => void): void {
+    this.eventQueue.push({ body, receiver } );
+    this.processQueuedEvents();
+  }
+
+  private processQueuedEvents() {
+    if (this.eventMode !== EventProcessingMode.Waiting) {
+      return;
+    }
+
+    const eventHandler = this.eventQueue.shift();
+    if (eventHandler === undefined) {
+      return;
+    }
+    const { body, receiver } = eventHandler;
+    this.eventMode = EventProcessingMode.Running;
+    this.continuationsRTS.runtime(body, (result) => {
+      this.eventMode = EventProcessingMode.Waiting;
+      receiver(result);
+      this.processQueuedEvents();
+    });
+  }
 }